Create abstract 'TaskManagerTest' class.
Move all same tests of managers to one abstract class.
